+++ src/screens/Project/TaskManagement/styles/ConfirmationDialogs.scss
@@ -8,6 +8,7 @@
     line-height: 32px;
     text-align: left;
     margin-bottom: 16px;
+    margin-top: 10px;
   }
 
   p {
